% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{fertilisExmpl}
\alias{fertilisExmpl}
\title{Example of nitogen fertilisation input}
\format{
A data frame of harvest dates with 8 row and 2 variables:
\describe{
  \item{Dates}{Fertilisation dates in julian days}
  \item{kgN_ha}{fertilisation amount in kg/ha of equivalent mineral nitrogen}

}
}
\usage{
fertilisExmpl
}
\description{
A dataset containing example nitrogen fertilisation dates and amounts
}
\keyword{datasets}
