% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{harvestExmpl}
\alias{harvestExmpl}
\title{Example of harvest input}
\format{
A named vector of harvest dates in julian days
}
\usage{
harvestExmpl
}
\description{
A dataset containing example harvest dates
}
\keyword{datasets}
