% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{lingraOutputExmpl}
\alias{lingraOutputExmpl}
\title{Example of output form Lingranr() function}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 365 rows and 423 columns.
}
\usage{
lingraOutputExmpl
}
\description{
A dataset the the full output of Lingranr function. This is equivalent to
the "Computations" sheed int the LINGRA-N Tool.
See ?Lingranr() for details.
}
\keyword{datasets}
