% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaModel.R
\name{Nchange}
\alias{Nchange}
\title{Nitrogen fertilisation effect on yield}
\usage{
Nchange(yield, mng, N)
}
\arguments{
\item{yield}{Numeric. Current dry matter yield (e.g. in kg/ha or t/ha)}

\item{mng}{Character. Current management, either "TG" (Temporary Grassland),
"PG" (Permanent Grassland), "RG" (Rough Grazing).}

\item{N}{Numeric value of nitrogen fertilisation in kg/ha.}
}
\value{
Numeric. The new yield after modifying nitrogen fertilsation in the
same units as \code{yield}.
}
\description{
This function modifies the yield output from the \link{metamodel}
functions according user-defined changes in nitrogen fertilisation.
}
\details{
This function is based on the four-parameter rational equation
proposed by Morrison et al. (1980) as parameterised by Qi et al. (2018)
for the UK. See \link{metamodel} for default nitrogen inputs of 
grassland types.
}
\examples{
# Yield of temporary grassland:
# (Metamodel's default for temporary grassland is 300 kg/ha N fertilisation)
tgy <- tg(awc = 37.9,
          pmam = 105,
          pjj = 126.9,
          pas = 95.2,
          rmam = 1247.7,
          ras = 786.8,
          tmam = 9.8,
          tjj = 15.8,
          tas = 15.6)

# Yield when fertilised with 500 kg/ha
tgyN <- Nchange(tgy, "TG", 500)
print(c(tgy = tgy, tgyN = tgyN))

}
