% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFuctions.R
\name{plotdm}
\alias{plotdm}
\title{Plot dry matter production}
\usage{
plotdm(lo)
}
\arguments{
\item{lo}{Tibble as outputted from function Lingranr().
See ?lingranr() for details.}
}
\description{
Plots the dry matter production in a field
}
\examples{

data(lingraOutputExmpl)
plotdm(lingraOutputExmpl)

}
