% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFuctions.R
\name{plotgr}
\alias{plotgr}
\title{Plot growth rates}
\usage{
plotgr(lo)
}
\arguments{
\item{lo}{Tibble as outputted from function Lingranr()}
}
\description{
Plots the growth rates
}
\examples{

data(lingraOutputExmpl)
plotgr(lingraOutputExmpl)


}
