% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFuctions.R
\name{plotla}
\alias{plotla}
\title{Plot leaf area}
\usage{
plotla(lo)
}
\arguments{
\item{lo}{Tibble as outputted from function Lingranr()}
}
\description{
Plots the development of the photosynthetic surface in the field.
}
\examples{

data(lingraOutputExmpl)
plotla(lingraOutputExmpl)

}
