% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFuctions.R
\name{plotn}
\alias{plotn}
\title{Plot nitrogen balance}
\usage{
plotn(lo)
}
\arguments{
\item{lo}{Tibble as outputted from function Lingranr().
See ?lingranr() for details.}
}
\description{
Plots the nitrogen balance between soil and grass
}
\examples{

data(lingraOutputExmpl)
plotn(lingraOutputExmpl)


}
