% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFuctions.R
\name{plotw}
\alias{plotw}
\title{Plot weather}
\usage{
plotw(w)
}
\arguments{
\item{w}{Weather data as inputted in Lingranr().
See ?lingranr() for details.}
}
\description{
Plots the weather input for Lingranr()
}
\examples{

data(weatherExmpl)
plotw(weatherExmpl[1:365,])


}
