% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{soilExmpl}
\alias{soilExmpl}
\title{Example of soil input}
\format{
A data frame with 1 row and 4 variables:
\describe{
  \item{sdpt}{soil depth [mm]}
  \item{mcfc}{moisture content at field capacity [mm3/mm3]}
  \item{mcsat}{moisture content at saturation [mm3/mm3]}
  \item{mcwp}{moisture content at wielting point [mm3/mm3]}
  ...
}
}
\usage{
soilExmpl
}
\description{
A dataset containing example soil input data for function Lingranr()
for one location
}
\keyword{datasets}
