% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaModel.R
\name{tg}
\alias{tg}
\alias{metamodel}
\alias{pg}
\alias{rg}
\title{Run metamodel}
\usage{
tg(awc, pmam, pjj, pas, rmam, ras, tmam, tjj, tas)

pg(awc, pamj, pjas, ramj, rjas, tmam, tjas)

rg(awc, pamj, pjas, ramj, rjas, tmam, tjas)
}
\arguments{
\item{awc}{Numeric. Available water content}

\item{pmam}{Cumulative precipitation for the period March-April-May}

\item{pjj}{Cumulative precipitation for the period June-July}

\item{pas}{Cumulative precipitation for the period August-September}

\item{rmam}{Cumulative global radiation for the period March-April-May}

\item{ras}{Cumulative global radiation for the period August-September}

\item{tmam}{Mean air temperature for the period March-April-June}

\item{tjj}{Mean air temperature for the period June-July}

\item{tas}{Mean air temperature for the period August-September}

\item{pamj}{Cumulative precipitation for the period April-May-June}

\item{pjas}{Cumulative precipitation for the period July-August-September}

\item{ramj}{Cumulative global radiation for the period April-May-June}

\item{rjas}{Cumulative global radiation for the period July-August-September}

\item{tjas}{Mean air temperature for the period June-August-September}
}
\value{
Numeric value of dry matter production in t/ha.
}
\description{
These functions run the metamodel described in Qi et al. 2018 for
grassland productivity on temporary (\code{tg()}), permanent (\code{pg()})
and semi-natural (rough grazing, \code{rg()}) grassland.
}
\details{
"Temporary grasslands
are the most productive, often consisting of frequently resown
perennial ryegrasses (Lolium perenne) and are assumed to receive
annual N application rate of 300 kg N ha−1.
Permanent grasslands
consist of a mixture of sown and indigenous grasses and
legumes; they are of intermediate productivity and are assumed to receive
annual N applications of 150 kg N ha−1.
The extensively used rough grazing
are diverse semi-natural grasslands containing various herbaceous species,
receive no synthetic N and are areas of low productivity." (Qi et al. 2018)
}
\examples{
# Yield of temporary grassland:
tgy <- tg(awc = 37.9,
          pmam = 105,
          pjj = 126.9,
          pas = 95.2,
          rmam = 1247.7,
          ras = 786.8,
          tmam = 9.8,
          tjj = 15.8,
          tas = 15.6)

}
