% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{weatherExmpl}
\alias{weatherExmpl}
\title{Example of weather input}
\format{
A data frame with 4017 row and 10 variables:
\describe{
  \item{Year}{Year}
  \item{DOY}{Julian day}
  \item{RAD}{Total solar radiation [MJ/m2*day]}
  \item{MINTMP}{Minimum daily tempterature ['C]}
  \item{MAXTMP}{Maximum daily tempterature ['C]}
  \item{VP}{Vapour pressure [kPa]}
  \item{WIND}{Mean daily wind speed [m/s]}
  \item{RAIN}{Total daily rainfall [mm]}
  \item{Country}{Country of  location}
  \item{Site}{Name of location}

}
}
\usage{
weatherExmpl
}
\description{
A dataset containing example weather input data for function Lingranr()
for one location and 11 years
}
\keyword{datasets}
